requirements.txt
Copystreamlit
panda
matplotlib
seaborn
Pillow
streamlit-option-menu